pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-devproxy"
version = "0.3.0"
description = "LLM development debug layer - every API call recorded, nothing lost. Reasoning tokens, alerts, export, streaming."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["llm", "openai", "anthropic", "gemini", "development", "debugging", "cache", "semantic-cache", "reasoning-tokens", "streaming", "cost-management"]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.20.0"]
gemini = ["google-generativeai>=0.5.0"]
semantic-local = ["sentence-transformers>=2.6.0"]
semantic-openai = ["openai>=1.0.0"]
proxy = ["fastapi>=0.110.0", "uvicorn>=0.29.0", "httpx>=0.27.0"]
web = ["fastapi>=0.110.0", "uvicorn>=0.29.0", "jinja2>=3.1.0"]
all = [
    "openai>=1.0.0",
    "anthropic>=0.20.0",
    "google-generativeai>=0.5.0",
    "sentence-transformers>=2.6.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "httpx>=0.27.0",
    "jinja2>=3.1.0",
]
dev = [
    "pytest>=7.0.0",
    "openai>=1.0.0",
    "anthropic>=0.20.0",
    "sentence-transformers>=2.6.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "httpx>=0.27.0",
    "jinja2>=3.1.0",
]

[project.scripts]
llm-devproxy = "llm_devproxy.cli.commands:main"

[project.urls]
Homepage = "https://github.com/asahik44/llm-devproxy"

[tool.hatch.build.targets.wheel]
packages = ["llm_devproxy"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
